requirements.txt
Flask==3.0.0
openai
python-dotenv